file: Split up `file` module
